Split api/types.rs into api/types/ directory
Organize ODBC type definitions into separate submodules:
- api/types/cdata_types.rs — CDataType enum, constants, type aliases
- api/types/types.rs — all other ODBC types, structs, and handle helpers
- api/types/mod.rs — re-exports both submodules

Revert error.rs back to api/error.rs since CDataType now lives in the
same api/ module and can access OdbcError directly.

Made-with: Cursor
